refactor(tui): extract StepRenderer
